WIP: Move taker/maker specific actors into their crates
