+++ daemon-tests/Cargo.toml
@@ -9,11 +9,13 @@ edition = "2021"
 anyhow = "1"
 async-trait = "0.1"
 daemon = { path = "../daemon" }
+maker = { path = "../maker" }
 mockall = "0.11"
 mockall_derive = "0.11"
 rand = "0.6"
 rust_decimal = "1.20"
 rust_decimal_macros = "1.20"
+taker = { path = "../taker" }
 time = "0.3"
 tokio = { version = "1", features = ["rt-multi-thread", "macros", "sync", "net"] }
 tokio-tasks = { path = "../tokio-tasks" }
