+++ daemon/Cargo.toml
@@ -39,7 +39,7 @@ tokio-tasks = { path = "../tokio-tasks" }
 tokio-util = { version = "0.6", features = ["codec"] }
 tracing = { version = "0.1" }
 uuid = { version = "0.8", features = ["serde", "v4"] }
-x25519-dalek = { version = "1.1" }
+x25519-dalek = "1"
 xtra = { version = "0.6" }
 xtra-bitmex-price-feed = { path = "../xtra-bitmex-price-feed" }
 xtra_productivity = { version = "0.1.0" }
