+++ maker/src/maker_cfd.rs
@@ -1,41 +1,42 @@
-use crate::cfd_actors;
-use crate::cfd_actors::insert_cfd_and_update_feed;
 use crate::collab_settlement_maker;
-use crate::command;
 use crate::maker_inc_connections;
-use crate::model::cfd::Cfd;
-use crate::model::cfd::CollaborativeSettlementCompleted;
-use crate::model::cfd::Order;
-use crate::model::cfd::OrderId;
-use crate::model::cfd::Origin;
-use crate::model::cfd::Role;
-use crate::model::cfd::RolloverCompleted;
-use crate::model::cfd::RolloverProposal;
-use crate::model::cfd::SettlementProposal;
-use crate::model::cfd::SetupCompleted;
-use crate::model::FundingRate;
-use crate::model::Identity;
-use crate::model::OpeningFee;
-use crate::model::Position;
-use crate::model::Price;
-use crate::model::TxFeeRate;
-use crate::model::Usd;
-use crate::monitor;
-use crate::oracle;
-use crate::process_manager;
-use crate::projection;
-use crate::projection::Update;
 use crate::rollover_maker;
 use crate::setup_maker;
-use crate::wallet;
-use crate::wire;
-use crate::wire::TakerToMaker;
 use anyhow::anyhow;
 use anyhow::bail;
 use anyhow::Context;
 use anyhow::Result;
 use async_trait::async_trait;
-use bdk::bitcoin::secp256k1::schnorrsig;
+use daemon::bdk::bitcoin::secp256k1::schnorrsig;
+use daemon::cfd_actors;
+use daemon::cfd_actors::insert_cfd_and_update_feed;
+use daemon::command;
+use daemon::model::cfd::Cfd;
+use daemon::model::cfd::CollaborativeSettlementCompleted;
+use daemon::model::cfd::Order;
+use daemon::model::cfd::OrderId;
+use daemon::model::cfd::Origin;
+use daemon::model::cfd::Role;
+use daemon::model::cfd::RolloverCompleted;
+use daemon::model::cfd::RolloverProposal;
+use daemon::model::cfd::SettlementProposal;
+use daemon::model::cfd::SetupCompleted;
+use daemon::model::FundingRate;
+use daemon::model::Identity;
+use daemon::model::OpeningFee;
+use daemon::model::Position;
+use daemon::model::Price;
+use daemon::model::TxFeeRate;
+use daemon::model::Usd;
+use daemon::monitor;
+use daemon::oracle;
+use daemon::process_manager;
+use daemon::projection;
+use daemon::projection::Update;
+use daemon::sqlx;
+use daemon::wallet;
+use daemon::wire;
+use daemon::wire::TakerToMaker;
 use std::collections::HashSet;
 use time::Duration;
 use tokio_tasks::Tasks;
