+++ maker/src/rollover_maker.rs
@@ -1,27 +1,27 @@
-use crate::command;
 use crate::maker_inc_connections;
 use crate::maker_inc_connections::RegisterRollover;
 use crate::maker_inc_connections::TakerMessage;
-use crate::model::cfd::Completed;
-use crate::model::cfd::Dlc;
-use crate::model::cfd::OrderId;
-use crate::model::cfd::Role;
-use crate::model::cfd::RolloverCompleted;
-use crate::model::FundingFee;
-use crate::model::FundingRate;
-use crate::model::Identity;
-use crate::model::TxFeeRate;
-use crate::oracle;
-use crate::oracle::GetAnnouncement;
-use crate::process_manager;
-use crate::schnorrsig;
-use crate::setup_contract;
-use crate::wire;
-use crate::wire::MakerToTaker;
-use crate::wire::RolloverMsg;
-use crate::Stopping;
 use anyhow::Context as _;
 use anyhow::Result;
+use daemon::command;
+use daemon::maia::secp256k1_zkp::schnorrsig;
+use daemon::model::cfd::Completed;
+use daemon::model::cfd::Dlc;
+use daemon::model::cfd::OrderId;
+use daemon::model::cfd::Role;
+use daemon::model::cfd::RolloverCompleted;
+use daemon::model::FundingFee;
+use daemon::model::FundingRate;
+use daemon::model::Identity;
+use daemon::model::TxFeeRate;
+use daemon::oracle;
+use daemon::oracle::GetAnnouncement;
+use daemon::process_manager;
+use daemon::setup_contract;
+use daemon::sqlx;
+use daemon::wire;
+use daemon::wire::MakerToTaker;
+use daemon::wire::RolloverMsg;
 use futures::channel::mpsc;
 use futures::channel::mpsc::UnboundedSender;
 use futures::future;
@@ -31,6 +31,7 @@ use xtra::prelude::MessageChannel;
 use xtra::Context;
 use xtra::KeepRunning;
 use xtra_productivity::xtra_productivity;
+use xtras::address_map::Stopping;
 
 /// Upon accepting Rollover maker sends the current estimated transaction fee and
 /// funding rate
