+++ maker/src/setup_maker.rs
@@ -1,28 +1,29 @@
-use crate::command;
 use crate::maker_inc_connections;
 use crate::maker_inc_connections::TakerMessage;
-use crate::model::cfd::Dlc;
-use crate::model::cfd::Order;
-use crate::model::cfd::OrderId;
-use crate::model::cfd::Role;
-use crate::model::cfd::SetupCompleted;
-use crate::model::Identity;
-use crate::model::Usd;
-use crate::oracle::Announcement;
-use crate::process_manager;
-use crate::setup_contract;
-use crate::wallet;
-use crate::wire;
-use crate::wire::MakerToTaker;
-use crate::wire::SetupMsg;
 use anyhow::Context;
 use anyhow::Result;
 use async_trait::async_trait;
+use daemon::command;
+use daemon::maia::secp256k1_zkp::schnorrsig;
+use daemon::model::cfd::Dlc;
+use daemon::model::cfd::Order;
+use daemon::model::cfd::OrderId;
+use daemon::model::cfd::Role;
+use daemon::model::cfd::SetupCompleted;
+use daemon::model::Identity;
+use daemon::model::Usd;
+use daemon::oracle::Announcement;
+use daemon::process_manager;
+use daemon::setup_contract;
+use daemon::sqlx;
+use daemon::wallet;
+use daemon::wire;
+use daemon::wire::MakerToTaker;
+use daemon::wire::SetupMsg;
 use futures::channel::mpsc;
 use futures::channel::mpsc::UnboundedSender;
 use futures::future;
 use futures::SinkExt;
-use maia::secp256k1_zkp::schnorrsig;
 use tokio_tasks::Tasks;
 use xtra::prelude::MessageChannel;
 use xtra::Address;
