+++ shared-bin/src/to_sse_event.rs
@@ -1,7 +1,4 @@
-use crate::ConnectionCloseReason::MakerVersionOutdated;
-use crate::ConnectionCloseReason::TakerVersionOutdated;
 use daemon::bdk::bitcoin::Amount;
-use daemon::connection;
 use daemon::model;
 use daemon::model::Identity;
 use daemon::model::Timestamp;
@@ -50,46 +47,6 @@ impl ToSseEvent for Option<model::WalletInfo> {
         });
 
         Event::json(&wallet_info).event("wallet")
-    }
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct ConnectionStatus {
-    online: bool,
-    connection_close_reason: Option<ConnectionCloseReason>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub enum ConnectionCloseReason {
-    MakerVersionOutdated,
-    TakerVersionOutdated,
-}
-
-impl ToSseEvent for connection::ConnectionStatus {
-    fn to_sse_event(&self) -> Event {
-        let connected = match self {
-            connection::ConnectionStatus::Online => ConnectionStatus {
-                online: true,
-                connection_close_reason: None,
-            },
-            connection::ConnectionStatus::Offline { reason } => ConnectionStatus {
-                online: false,
-                connection_close_reason: reason.as_ref().map(|g| match g {
-                    connection::ConnectionCloseReason::VersionMismatch {
-                        maker_version,
-                        taker_version,
-                    } => {
-                        if *maker_version < *taker_version {
-                            MakerVersionOutdated
-                        } else {
-                            TakerVersionOutdated
-                        }
-                    }
-                }),
-            },
-        };
-
-        Event::json(&connected).event("maker_status")
     }
 }
