+++ taker/src/auto_rollover.rs
@@ -1,15 +1,16 @@
-use crate::cfd_actors::load_cfd;
 use crate::connection;
-use crate::db;
-use crate::model::cfd::OrderId;
-use crate::oracle;
-use crate::process_manager;
 use crate::rollover_taker;
-use crate::try_continue;
 use anyhow::Context;
 use anyhow::Result;
 use async_trait::async_trait;
-use maia::secp256k1_zkp::schnorrsig;
+use daemon::cfd_actors::load_cfd;
+use daemon::db;
+use daemon::maia::secp256k1_zkp::schnorrsig;
+use daemon::model::cfd::OrderId;
+use daemon::oracle;
+use daemon::process_manager;
+use daemon::sqlx;
+use daemon::try_continue;
 use std::time::Duration;
 use time::OffsetDateTime;
 use tokio_tasks::Tasks;
