+++ taker/src/taker_cfd.rs
@@ -1,25 +1,26 @@
-use crate::cfd_actors;
-use crate::cfd_actors::insert_cfd_and_update_feed;
 use crate::collab_settlement_taker;
 use crate::connection;
-use crate::model::cfd::Cfd;
-use crate::model::cfd::Order;
-use crate::model::cfd::OrderId;
-use crate::model::cfd::Origin;
-use crate::model::cfd::Role;
-use crate::model::Identity;
-use crate::model::Position;
-use crate::model::Price;
-use crate::model::Usd;
-use crate::monitor;
-use crate::oracle;
-use crate::process_manager;
-use crate::projection;
 use crate::setup_taker;
-use crate::wallet;
 use anyhow::Context as _;
 use anyhow::Result;
-use bdk::bitcoin::secp256k1::schnorrsig;
+use daemon::bdk::bitcoin::secp256k1::schnorrsig;
+use daemon::cfd_actors;
+use daemon::cfd_actors::insert_cfd_and_update_feed;
+use daemon::model::cfd::Cfd;
+use daemon::model::cfd::Order;
+use daemon::model::cfd::OrderId;
+use daemon::model::cfd::Origin;
+use daemon::model::cfd::Role;
+use daemon::model::Identity;
+use daemon::model::Position;
+use daemon::model::Price;
+use daemon::model::Usd;
+use daemon::monitor;
+use daemon::oracle;
+use daemon::process_manager;
+use daemon::projection;
+use daemon::sqlx;
+use daemon::wallet;
 use tokio_tasks::Tasks;
 use xtra::prelude::*;
 use xtra::Actor as _;
